refactor: added requested change
